pyproject.toml
[project]
name = "eob_search"
version = "0.1.0"
description = "Default template for PDM package"
authors = [
    {name = "Fedor Krasnov", email = "fedor.krasnov@gmail.com"},
]
dependencies = ["fastapi>=0.116.0", "uvicorn>=0.35.0", "elasticsearch>=9.0.2", "sentence-transformers>=5.0.0", "lightgbm>=4.6.0", "scikit-learn>=1.7.0", "numpy", "pydantic>=2.11.7", "aiohttp>=3.12.14", "PyYAML>=6.0.2"]
requires-python = "==3.12.*"
readme = "README.md"
license = {text = "MIT"}

[tool.pytest.ini_options]
pythonpath = ["."]
minversion = "6.0"
addopts = "-ra -q --cov=app --cov-report=term-missing"

[tool.ruff]
exclude = [".venv", "__init__.py"]

[tool.pdm]
distribution = false
pythonpath = [".", ".."]

[tool.pdm.scripts]
test = "pytest tests/"
measure_metrics = "python -m scripts.measure_metrics"
create_es_index = "python -m scripts.create_es_index"
train_reranker = "python -m scripts.train_reranker"


[dependency-groups]
test= [
    "pytest",
    "pytest-cov",
    "pre-commit",
    "ruff>=0.12.0",
    "pytest-asyncio",
    "respx", 
    "numpy", 
]
metrics = [
    "requests",
    "scikit-learn",
    "numpy",
]
